Moved win screen to the draw gameplay function
